#17: Added tests and postgres for travis

+++ src/main/scala/no/gdl/bookapi/model/api/Book.scala
@@ -17,7 +17,7 @@ import org.scalatra.swagger.runtime.annotations.ApiModelProperty
 import scala.annotation.meta.field
 
 @ApiModel(description = "Information about the licensing of the book")
-case class License(@(ApiModelProperty@field)(description = "The id of the license") id: Int,
+case class License(@(ApiModelProperty@field)(description = "The id of the license") id: Long,
                    @(ApiModelProperty@field)(description = "The revision of the license") revision: Int,
                    @(ApiModelProperty@field)(description = "The name of the license") name: String,
                    @(ApiModelProperty@field)(description = "Description of the license") description: Option[String],
